refactor: unify integration tests under a single shared base
- Add shared IntegrationTest base with app, clock, and DB cleanup lifecycle hooks
- Migrate existing integration tests to extend the shared base
- Move UsersContractIntegrationTest to server.api as global ContractIntegrationTest
- Extract createValidJwt and getValueFromCookie helpers into shared Helpers
