+++ backend/api/src/test/java/com/anibalxyz/features/auth/AuthRoutesIntegrationTest.java
@@ -3,7 +3,7 @@ package com.anibalxyz.features.auth;
 import static com.anibalxyz.features.auth.api.AuthController.REFRESH_TOKEN_COOKIE;
 import static com.anibalxyz.shared.Constants.Auth.VALID_REFRESH_TOKEN;
 import static com.anibalxyz.shared.Constants.Users.*;
-import static com.anibalxyz.shared.Helpers.cleanDatabase;
+import static com.anibalxyz.shared.Helpers.getValueFromCookie;
 import static com.anibalxyz.shared.Helpers.persistUser;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -20,65 +20,35 @@ import com.anibalxyz.features.common.api.out.response.error.ErrorResponse;
 import com.anibalxyz.features.users.domain.Email;
 import com.anibalxyz.features.users.domain.User;
 import com.anibalxyz.features.users.domain.error.UserDomainError;
-import com.anibalxyz.server.Application;
 import com.anibalxyz.server.api.ErrorMapper;
 import com.anibalxyz.server.api.ErrorResult;
 import com.anibalxyz.server.api.InfrastructureErrorMapper;
 import com.anibalxyz.server.config.environment.AppEnvironmentSource;
 import com.anibalxyz.shared.Constants;
-import com.anibalxyz.shared.HttpRequest;
-import com.anibalxyz.shared.MutableClock;
+import com.anibalxyz.shared.IntegrationTest;
 import com.anibalxyz.shared.ResultAsserts;
 import io.javalin.http.UnauthorizedResponse;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
 import java.time.*;
 import java.time.temporal.TemporalAdjusters;
 import java.util.Date;
 import java.util.Map;
-import okhttp3.OkHttpClient;
 import okhttp3.Response;
 import org.junit.jupiter.api.*;
-import tools.jackson.databind.ObjectMapper;
 
 @DisplayName("Tests for AuthRoutes")
-public class AuthRoutesIntegrationTest {
-  // Tuesday 10:00
-  private static final ZonedDateTime FIXED_NOW =
-      LocalDateTime.of(2026, 4, 21, 10, 0).atZone(ZoneId.of("America/Montevideo"));
-  private static final MutableClock testClock =
-      new MutableClock(FIXED_NOW.toInstant(), FIXED_NOW.getZone());
+public class AuthRoutesIntegrationTest extends IntegrationTest {
   private static final Instant SATURDAY_MIDDAY =
       FIXED_NOW.with(TemporalAdjusters.next(DayOfWeek.SATURDAY)).with(LocalTime.NOON).toInstant();
   private static final Instant MAINTENANCE_START =
       FIXED_NOW.with(TemporalAdjusters.next(DayOfWeek.MONDAY)).with(LocalTime.of(8, 0)).toInstant();
-  private static Application app;
-  private static EntityManagerFactory emf;
-  private static HttpRequest http;
   private static JwtService jwtService;
   private static AppEnvironmentSource env;
   private static RefreshTokenService refreshTokenService;
-  private EntityManager em;
 
   @BeforeAll
   public static void setup() {
-    Constants.init();
-    app = Application.create(Constants.APP_CONFIG, testClock);
-    app.start(0);
-
-    String baseUrl = app.javalin().jettyServer().server().getURI().toString() + "api";
-    emf = app.persistenceManager().emf();
-    ObjectMapper objectMapper = new ObjectMapper();
-
-    http = new HttpRequest(objectMapper, new OkHttpClient(), baseUrl);
-
     jwtService = new JwtService(Constants.APP_CONFIG.env(), testClock);
     env = Constants.APP_CONFIG.env();
-  }
-
-  @AfterAll
-  public static void shutdown() {
-    app.stop();
   }
 
   private static void validateJwt(String accessToken, Integer id) {
@@ -100,27 +70,9 @@ public class AuthRoutesIntegrationTest {
   }
 
   @BeforeEach
-  public void openEntityManager() {
-    em = emf.createEntityManager();
-    cleanDatabase(em);
-  }
-
-  @BeforeEach
-  public void di() {
+  public void deps() {
     var refreshTokenRepository = new JpaRefreshTokenRepository(() -> em);
     refreshTokenService = new RefreshTokenService(refreshTokenRepository);
-  }
-
-  @BeforeEach
-  public void resetClock() {
-    testClock.resetTo(FIXED_NOW.toInstant());
-  }
-
-  @AfterEach
-  public void closeEntityManager() {
-    if (em.isOpen()) {
-      em.close();
-    }
   }
 
   private LoginResult loginUser(String email, String password) {
@@ -131,19 +83,6 @@ public class AuthRoutesIntegrationTest {
     return new LoginResult(
         authResponse.accessToken(),
         getValueFromCookie(loginResponse.header("Set-Cookie"), REFRESH_TOKEN_COOKIE));
-  }
-
-  private static String getValueFromCookie(String cookie, String key) {
-    if (cookie == null) {
-      return null;
-    }
-    for (String cookiePart : cookie.split(";")) {
-      String[] parts = cookiePart.trim().split("=");
-      if (parts.length > 0 && parts[0].equals(key)) {
-        return parts.length > 1 ? parts[1] : "";
-      }
-    }
-    return null;
   }
 
   @Test
